fixed data type mismatch
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@21999 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/net/vos/voslightmapcache.cpp
@@ -35,8 +35,8 @@ const char* LightmapCache::GetCurrentScope () const
     return "";
 }
 
-bool LightmapCache::CacheData (const void* data, unsigned long size, const char* type,
-                               const char* scope, unsigned int id)
+bool LightmapCache::CacheData (const void* data, size_t size, const char* type,
+                               const char* scope, uint32 id)
 {
     LOG("lightmapcache", 3, "caching " << size << " bytes");
     property->replace(std::string((char*)data, size));
